Storagev4/requirements.txt
a2wsgi==1.10.0
anyio==4.2.0
APScheduler==3.10.4
argcomplete==1.10.3
asgiref==3.7.2
attrs==23.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
PyMySQL==1.1.0
click==8.1.3
colorama==0.4.6
connexion==3.0.5
fastapi==0.109.0
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
httpcore==1.0.2
pykafka==2.8.0
httpx==0.26.0
idna==3.6
inflection==0.5.1
Jinja2==3.1.2
jsonschema==4.21.1
mysql-connector-python==8.0.32 
colorlog==6.8.2
packaging==23.0
prompt-toolkit==3.0.43
psutil==5.9.7
pydantic==2.5.3
Pygments==2.17.2
python-dateutil==2.8.2
pytz==2023.4
requests==2.31.0
SQLAlchemy==2.0.25
starlette==0.32.0.post1
swagger_ui_bundle==1.1.0
typing_extensions==4.9.0
urllib3==2.1.0
uvicorn==0.26.0
Werkzeug==3.0.1